Add local routes to source routed tables
This fixes an issue where responses to pings and inbound connections
from LAN devices would get sent to the default gateway. This actually
"worked" in a lot of setups. The fix is to add the LAN route to the
policy-based routing tables too. This also changes the order of calls to
"ip" so that the rules come first. This feels better since the local
routes reference the tables. It appears unnecessary in practice, since
Linux lets you operate on routing tables without rules that refer to
them.