Add helper to force WiFi scan in AP mode
For WiFi configuration, it is nice to start a WiFi module up in AP mode
so that users can connect to it and select a WiFi network for the device
to use. Scanning in AP mode is problematic since the WiFi module needs
to look at other frequencies than the one that it's on to see if there
are APs around. Currently the wpa_supplicant doesn't support doing this.
To work around that, this includes a small C program that sends the
right message through the nl80211 netlink interface. Discovered APs go
through the wpa_supplicant and are reported like normal.