Handle WiFi drivers w/o separate P2P interfaces
Not all WiFi drivers expose a separate P2P control interface. This
refactors the control file logic so that it tries both the normal
control interface and the P2P interface when the driver is in AP mode.

This fixes an issue with a USB WiFi module that doesn't have this
interface. The Raspberry Pi, for example, has both the P2P and normal
control interface, but only the P2P interface is used when in AP mode.