Report WiFi access points as a list
This is a non-backwards compatible update that changes how access points
are reported to users.

The new way is to report a list of access points. At one point reporting
maps made some sense, but now it seems more of an annoyance to code that
uses vintage_net. This should clean up the map->list conversions that
are currently being done by users (after they update their code).