Patch race in unit tests
The race is that it takes a short period of time for vintage_net to
detect network interfaces and the detection process is asynchronous. The
unit test that fails is focused on the synchronous call to list
interfaces and if that gets called too quickly after the app is started,
it won't list the loopback interface. Add a short sleep to fix.