Use ip rather than ifupdown; refactor network config
Switching from ifupdown to calling ip directly makes it possible to
supervise udhcpc properly. Previously, it was being started outside of
supervision from ifup. That made it awkward to kill and resulted in
udhcpc processes stacking up under certain conditions.

As part of this refactor to use IP, many changes were made to
configuration processing. The most externally visibly obvious changes
are the normalization of IP addresses to tuple form and that WiFi
networks are all under the `:networks` key now.