Start the wpa_supplicant from WpaSupplicant
There's a forced ordering between the C wpa_supplicant and the Elixir
WpaSupplicant. This has a big benefit in grouping their crashes together
and in the ability to remove old control files. The latter was causing
ATTACH timeouts and WpaSupplicant use that would fail after 5 seconds
(the default timeout).

Fixes #116