Rename the "disconnected interface" to avoid it existing
Depending on which order ex_unit runs the Internet and LAN connectivity
tests, one or the other will break due to the "disconnected interface"
not sending a notification since the user is supposed to already know
that it's disconnected. There was no bug. It's just that testing the
same interface had the effect of the tests affecting each other.