Add option to turn off configuration persistence
This allows users to try out a configuration or temporarily override a
configuration. If anything goes wrong and the
device gets rebooted, the old configuration will be applied.

This is needed for network configuration wizards. For example, the
`vintage_net_wizard` switches WiFi adapters to AP mode for
configuration. Consider what happens if configuration mode were entered
accidentally and out of confusion a user reboots the device. With
persistence of AP mode, the device will still not work after a reboot.
However, if the AP mode isn't persisted, then rebooting will fix the
situation.