Make max interface count part of API
This unhides an internal limitation of how many interfaces that
VintageNet supports. Linux provides one constraint that's pretty high.
VintageNet's use of the `ip rule` puts a fixable limit of about 150
network interfaces. It is, however, nice to reduce the limit to
something more realistic and small so avoid noise when cleaning up
tables. This sets it to 8 network interfaces per device. That seems
more than sufficient for most Nerves users.