Fix missing address update on dhcp restart
If the DHCP client was restarted (for any reason), it first reports that
the network interface should be deconfigured and then reports the new
(likely the same) lease. This works. However, the deconfig/renew ends up
clearing multicast registrations on the IP address being
deconfig/renew'd. This broke mDNS since it was listening on the IP
address.

To fix this, code is added to clear out all IPv4 addresses on deconfig.
This propogates the "glitch" in IP network connectivity throughout and
as soon as it could possibly happen. When the IP address comes back via
the renew, `mdns_lite` will register a listener and all will work.

It should be noted that the GenServer crash that caused this bug to
happen has been fixed. One way of reproducing is to back out
bc8073a5d694de3da1cd3169cb1002fc9ea9015f.