Split resolv.conf generation for better testing
This pulls the pure part of creating resolv.conf contents into its own
module. This allows for easier testing of future additions that will
involve more complicated manipulations of the file.

A side effect of this was to fix some sloppiness around IP address
handling. This isn't a big deal now, but it should make IPv6 support
easier to handle later.