Force the call to Technology.normalize/1
This fixes an issue where the Technology documentation says that
VintageNet calls the normalize/1 function, but it actually doesn't. All
of the existing to_raw_config/3 functions called it and that masked the
lack of a call. This fixes the issue and adds a unit test to avoid it
happening again.