Fix wrong IP address being used for ppp links
Linux reports the local IP address of a ppp link using the `:local` key
rather than the `:address` key. The latter key is for the remote side of
the link. For the standard "broadcast" network links, both `:local` and
`:address` are set the same so it didn't matter.

This fixes the issue, adds a test and leaves a note since this seems
unintuitive.