Support waiting for multiple interfaces
This is an API breaking change to fix an issue where technologies needed
to either wait for multiple interfaces to be present or wait for an
interface that was different from the one that it managed.

For example, cellular modems use PPP interfaces (e.g., ppp0), but that's
an interface that they bring up. It's common that they wait for `wwan0`
to exist first.

Another example are bridges that need to wait for the networks they're
supposed to bridge to become present before they're brought up.

All technologies need to change the RawConfig structs they generate to
use this feature. You will receive a compiler error, so it should be
hard to miss. First, update your `mix` dependencies to require
`:vintage_net, "~> 0.8.0"` to force the version update. Then whereever
you're creating the RawConfig, if you had `require_interface: false`,
change that to `required_ifnames: []`. If you did nothing (most
technologies), then you'll want to require the interface that you're
using. I.e., `required_ifnames: [ifname]`.