Remove ifupdown since they're no longer used
When we started, we used to use ifupdown since they made some
configuration so easy. However, we haven't used them in a really long
time and it seems unlikely that we'd ever go back.

This comment removes them to avoid giving the false impression that
anyone may need them to use `vintage_net_*`