Return errors on missing commands rather than raise
This makes it possible for the caller to decide whether a missing
command should crash. This is different from the semantics of
System.cmd/3 and MuonTrap.cmd/3, but it's so convenient so long as we
don't run into a command that does useful work and returns failure exit
statuses. To help prevent that, the exit status on missing command is
256 which should be unusual.

This also adds tests and documentation for posterity.