Don't fail completely when PowerManager.init/1 raises
When `PowerManager.init/1` implementations fail, supervision restarts
GenServers like expected. However, `init/1` really isn't supposed to
fail and when it does, it has the effect of propogating failures all the
way up which eventually takes down VintageNet. While this is almost
certainly a programming error, the result is particularly harsh.

To better support debugging this issue, this commit logs the error and
disables the power manager.

The downside to this is that if there's a transient that would have been
fixed by a restart, that no longer happens and perhaps the network won't
work.  While this isn't good, devices that require network connections
to function need to have a "watchdog" that verifies that there's a
network connection anyway. That logic should catch this.