Replace to_elixir code with beam_notify
BEAMNotify is better tested and used elsewhere, so this lets us delete
the code. The main change is that the we can't do the symlink trick to
use argv0 to differentiate calls. That's ok since udhcpc and udhcpd are
pretty easy to differentiate by looking at their arguments.

All of the unit tests have been updated. The tests for the C code were
converted to exercise the dispatcher function.

Since the word "to_elixir" doesn't make sense any more, this commit
renames quite a few modules, but doesn't change anything else
substantial.