Lack of IPv4 doesn't mean disconnected
The Internet connectivity checker would declare that the network was
disconnected if an IPv4 address wasn't available. This is untrue for a
couple reasons:

1. IPv6 addresses are available
2. DHCP and layer 2 protocols could still be used

This also had a weird side effect where interfaces would initially come
up (`lan` connectivity), then go disconnected when an IPv4 address
wasn't found, and then come back when an IPv4 address was found.

Fixes #268