Support >1 internet host to check; refactor to be easier to test
This adds the `:internet_host_list` configuration key and deprecates
`:internet_host`. The new key, `:internet_host_list`, takes a list of IP
address, port tuples and attempts to connect to each one of them in
succession to establish whether the internet is available on a network
interface..

This fixes #77. It also lets you specify the TCP port number to check so
that it's possible to test a connection to the service being run. As
before, once the connection is made, it's closed since that's good
enough for determining connectivity.

As part of this change, there was quite a bit of untested code and the
addition of the new code made me worried that it might break easily.
This pulls much of the side effect-free code out so that it can be
tested.

Nothing big was found when adding tests. The main changes were to
bulletproof functions being called with values that could theoretically
lead to harmless, but unintended behavior. For example, if the interface
was removed and this was caught when trying to connect, then a retry
timeout would be set even though the intention was for the interface to
be marked disconnected and have no timeout.