Allow get_configuration to return a default value
Sometimes when changing to different network interfaces some systems
may or may not have a configuration for the either the new interface or
the old one. Changing `get_configuration` to return a default value (or
nil by default) will allow users to check if a configuration exists or
not for the interface.

Also, providing the `get_configuration!` function keeps the current
behavior in the library.

The reason to need this over `configured_interfaces/1` is because that
function considers configurations with `VintageNet.Technology.Null` to
not be configured and will be filtered out.