Support resetting an interface to its defaults
This adds `VintageNet.reset_to_defaults/1` to reset a configuration to
whatever it would be on a factory reset. This is useful if you no longer
need to override a configuration and would like it to use the built-in
defaults. For example, if you're debugging a device and changing the
network, you can reset it to the defaults with out manually erasing the
persisted configuration.