Harden app env checks; silently add default_config
This funnels the call to get configs through one place. There was a
missed call in `vintage_net.ex` that would have crashed if the
environment was bad. This also adds a few unit tests to make sure that
the invalid config filter works and logs a warning.

Finally, this silently adds the `:default_config` key so that we can
switch the configuration key to it in a future major release without
breaking people who back off to old `:vintage_net` releases for testing
(probably just me). The intention of `:default_config` is to reduce the
confusion of why the network configuration doesn't change when you
change the application config. I.e., it's not a configuration, it's a
"default" configuration that's only used if the interface is available
and nothing has been configured on top of it.