Add duration field in property metadata
This adds the field `:duration` to the metadata dispatched when a
property changes. This will be useful for consuming applications to
records metrics on how long a property was one value before it changed
to the new value. Otherwise, the consuming application will need to keep
its own time tracking implementation. This is because the consuming
application will be notified of some change and thus the time stamp of
the old value has been replaced with the time stamp of the new value and
the old time stamp will be lost.