Internet connectivity checker: remove unfixable transition
If a "ping" returned that there wasn't a network interface to ping, this
would lead to the interface being marked disconnected with no way of
escaping if it was a transient error.

The original assumption was that if ping return this that it was just a
matter of time before the interface went down. However, if there's a
transient or unfortunate ordering of events and this happened, then the
checker would stop polling and the interface could never be marked as
internet-connected.