Add internet connectivity inspector
This is a new module that can determine whether a network interface is
connected to the internet by watching traffic on TCP sockets known to
Erlang/OTP. If a device is using long-lived TCP connections (like for
MQTT) that send keepalive messages, then it can declare
internet-connectivity just by watching to the tx and rx counters on the
long-lived connection. This helps reduce use of metered connections and
benefits devices behind restrictive firewalls that drop the pings.

This commit only adds the module. It is hooked up in a subsequent commit.