Move check handling logic into its own module
This also simplifies what happens when the connectivity check fails. All
errors are handled the same and there are no shortcuts to `:lan` even if
it looks like the interface is removed. The expectation is that the
interface will be properly removed shortly if it's for real and
everything else is suspect.

Moving the check handling logic into its own module is intended to make
it easier to adjust it and to simplify addng the Inspector check to the
InternetChecker.