Use the connectivity inspector as a first internet check
This majorly reduces internet traffic for checking status on those
network interfaces with long-lived TCP connections. Basically, if a TCP
connection on the interface that's to connecting to a host on a
different LAN has sent and received bytes since the last internet check,
then the check is skipped. Internet is available. AWS IoT and NervesHub
connections are sufficient since their keepalive messages will satisfy
this check.