Fix internet connectivity integration tests
The connectivity checker is pickier now on whether the interface is up
or down so the unit tests can't take as many shortcuts.

This also fixes an issue where `ifup` would be called and then a
connectivity check done immediately afterwards. The intended behavior is
for the CheckLogic to decide rather than this to be assumed. The result
of this is that the check will be done 500 ms later. This seems good
since it's short and 500 ms allows for a little time for the interface
to settle after it just declared that it was up.