Move route metric calculation and remove custom priorities
This add docs and sets up the possibility of completely overriding route
metric prioritization. Going through the default calculation is
nontrivial due to a lack of information. It's way easier to just replace
the whole function with exactly what you want and that's the goal here.
As a result, the ability to dynamically set interface prioritization
order goes away. That's nice sense it deletes a bit of code that
probably wasn't used and was too inflexible anyway.

This is a backwards incompatible change for anyone setting an interface
prioritization table. If you don't use that, then this PR just moves
code around.