Support complete replacement of metric calculation
This adds a new key, `:route_metric_fun`, that let's users get rid of
the default metric calculator and use their own. The default calculator
probably works for most users, but once you find yourself outside of
what works with the defaults, it's probably easiest to code a solution
that exactly works with your setup rather than try to encourage the
default calculator to work for you.