Don't downgrade connection status on route change
The original logic would force the connection status to `:lan` if the IP
address or route changed. The logic was that the Internet might not be
reachible, so let the connectivity checker do it's thing and upgrade the
connection.

This had a couple undesirable side effects:

1. The interface preference would bounce when routes changed even if the route
   change resulted in the status quo.
2. There was a race condition between the connectivity checker. If it
   set "internet" right before the route change, then you'd have to wait
   until the next check before it would be restored to "internet".
3. Not all connectivity checkers were as persistent at setting
   "internet", so it could be degraded for a really long time.

This change keeps the interface connectivity the same or upgrades it to
"lan" on route changes. This basically changes the behavior from
pessimistic (the interface isn't internet-connected, so prove it) to
optimistic (the interface is still connected, and if that's wrong, let
me know).