Allow absolute path executables for commands
VintageNet currently expects the executable of a command
in the RawConfig to be relative in the PATH. However, some cases
you might need to use a binary compiled with your library (i.e
`vintage_net_wireguard` and meshing with `vintage_net_wifi`) and
want to provide the absolute path to an executable. This change
allows you do to that.

You might say "Well Jon, why don't you just edit the PATH?" -
Good question. The answer is that VintageNet sets a default path
which covers most of the cases. Expecting an altered PATH from
a library that depends on VintageNet means the user of the library
needs to be expected to set `config :vintage_net, :path` in their
library for the library which depends on VintageNet. Instead its
easier to do a sanity check on the "command" and consider it an
absolute path if it starts with `/`