Extract PropertyTable to its own library
*Backwards incompatible - please read*

PropertyTable is now its own library and it has been simplified. While
some API affecting changes to PropertyTable were made in the library to
fix bugs, improve performance and make it more generic, this preserves
the PropertyTable semantics as much as possible to avoid application
code needing to change. I don't believe any user code should need to
change.

The backwards incompatible part of this change is that all references to
`VintageNet.PropertyTable` need to be changed to `PropertyTable`. This
shouldn't affect most code, but does affect `vintage_net_wifi`.

This fixes #352.