Expose current name server list
This adds a new property, `"name_servers"`, so that VintageNet users can
easily get the list of name servers. Each name server is in a map with
its address and where it was defined. It's possible that port numbers
and other name server information is exposed in the future and that's
why this isn't a simple list of IP addresses.

Other changes in this commit:

1. VintageNet spells name server as two words like Wikipedia does. Any
   one-word versions were changed for consistency.
2. Remove the state struct in `NameResolver` to be consistent with other
   Nerves code that has moved away from using state structs.
3. Use `start_supervised!/1` instead of raw calls to `start_link/2` in
   tests.
4. Change `"global"` to an atom for when a name server is provided in
   the VintageNet configuration. Now that the API is public, it seems
   better to use an atom rather than accidentally conflict with an
   ifname of `"global"` (if that ever happens).