Don't prevent network changes if can't save
If the filesystem is messed up for some reason and the network
configuration couldn't be saved, the network configuration change was
aborted. This could be worked around by not saving the configuration.
However, this seemed like the wrong way to degrade. This commit changes
the logic to allow the network configuration, but log an error. The
configuration will be lost on a reboot, but perhaps it will help someone
debug and fix the device that's having trouble.