Fix resolv.conf being saved outside of the test directory
`NameResolverTest` uses the `in_tmp/2` helper function to scope tests to
temporary directories for that specific test. It includes a call to `File.cd!/1`
and deals with the resolv.conf file generated relatively. However, using
`start_supervised!/2` may not be changed into the tmp test directory and
`fake_resolv.conf` files could be created in the root of the project.

This makes an absolute path to the resolv.conf when starting the NameResolver process
to prevent those files from being generated in the root