Elevate DHCP options to their own module
This is a large change that removes the use of unmodified string maps
for representing DHCP options. Now that DHCP options are intentionally
part of the public interface, the keys used are now atoms with values
that have documented types.

Additionally, this makes some functions and modules involved with DHCP
options more obviously private. They need to be public to be easily
tested, but they may change in the future.