Fix confusion with dhcpd subnet option
Busybox calls the DHCP subnet mask option field "subnet" when
configuring it. This is reflected in VintageNet's configuration options
for it since they match one for one.

This cause some confusion on whether the subnet or subnet mask should be
passed. This fixes issues with the documentation and adds `:netmask` as
an alias so that users can make their own code more clear if they want.