Handle all messages sent to InterfacesMonitor
This adds handling for when the interfaces monitor port application
exits. This shouldn't happen, but it has been seen and this logs an
error with the test so that it's easier to read when it happens.

This PR also adds handling for other messages just in case some funny
business is going on with sending messages to the InterfacesMonitor.