Normalize configs on load and save failure reasons
Previously, the configurations in the application environment and those
persisted on the filesystem were kept unnormalized until used. You could
see this by getting the configuration via the property table or via
VintageNet.info.

This causes a couple problems. First, it was confusing since you'd get
different views of the configuration depending on whether it was set at
compile-time or run-time or restored. The second issue is that errors
due to run-time evaluation could be deferred for when you'd find out
about them.

This change normalizes the configurations on start and adds a reason
field to the `Null` configuration. That way if you miss or don't see the
error in the logs, you'll get a big message for why the interface wasn't
configured when checking its configuration.

This change also addresses a minor problem of why `Null` configurations
are set on interfaces since it can be set automatically on errors. If
the user did it themselves via `VintageNet.deconfigure/2`, this makes it
unambiguous.